moving some functions into common
