+++ src/whiteflag/wf_codec/common.rs
@@ -1,4 +1,17 @@
-use super::constants::BYTE;
+use super::constants::*;
+
+//https://github.com/WhiteflagProtocol/whiteflag-java/blob/57db4b6963a4a7913afdeb596e7ce11d46d9d93b/src/main/java/org/whiteflagprotocol/java/core/WfBinaryBuffer.java#L299
+pub fn to_hex(data: &Vec<u8>) -> String {
+    data.iter().flat_map(|b| convert_byte_to_hex(*b)).collect()
+}
+
+fn convert_byte_to_hex(byte: u8) -> [char; 2] {
+    let byte_u32 = byte as u32;
+    let c1 = std::char::from_digit((byte_u32 >> QUADBIT) & 0xF, HEXRADIX as u32)
+        .expect("encoding failed");
+    let c2 = std::char::from_digit(byte_u32 & 0xF, HEXRADIX as u32).expect("encoding failed");
+    [c1, c2]
+}
 
 /**
  * removes characters from string that are invalid in hexadecimal format
